refactor(types): break the four import cycles by moving types to their owners refs #281
madge reported four cycles against rule 28's zero. All four were a lower
layer importing a type back from a higher one:

- api/types.ts pulled EventFilters from api/events.ts. The interface now
  sits in types.ts with the other request shapes.
- lib/format-date-time.ts pulled DateFormatPreset and TimeFormatPreset
  from stores/settings. It already owns FormatSettings, which is built
  from both, so the presets belong there.
- lib/assistant/types.ts and lib/event/thumbnail-chain.ts pulled
  ThumbnailFallbackEntry from stores/settings. It now lives next to
  resolveFallbackFids, the code that interprets it.
- services/pushNotifications.ts pulled NotificationSettings and
  NotificationSource from stores/notifications. Both moved to
  types/notifications.ts, which already held NotificationMode and the
  rest of the domain's shared shapes.

Every original module re-exports what it used to declare, so no callsite
outside these six files changed. The cycles were type-only and erased at
build time under verbatimModuleSyntax, so this fixes the module graph
rather than a runtime defect: nothing stopped someone turning one of
those into a value import and making it real.

madge --circular is now clean.
