+++ app/src/types/notifications.ts
@@ -56,5 +56,30 @@ export type ConnectionState =
 
 export type NotificationMode = 'es' | 'direct';
 
+export interface MonitorNotificationConfig {
+    monitorId: number;
+    enabled: boolean;
+    checkInterval: number; // Seconds between checks (60, 120, etc.)
+}
+
+/** Per-profile notification configuration, persisted by stores/notifications. */
+export interface NotificationSettings {
+    enabled: boolean;
+    notificationMode: NotificationMode; // 'es' = Event Server websocket, 'direct' = ZM REST API
+    notificationId: number | null; // Server-side Notifications.Id (direct mode)
+    host: string; // Event server host (e.g., "zm.example.com")
+    port: number; // Event server port (default 9000)
+    ssl: boolean; // Use wss:// instead of ws://
+    allMonitors: boolean; // Receive notifications for all monitors (no filter sent to ES)
+    monitorFilters: MonitorNotificationConfig[]; // Per-monitor settings (used when allMonitors is false)
+    onlyDetectedEvents: boolean; // Only notify for events with object detection results (direct mode)
+    pollingInterval: number; // Seconds between event polls in direct mode (desktop)
+    showToasts: boolean; // Show toast notifications for events
+    playSound: boolean; // Play sound on notification
+    badgeCount: number; // Current unread count
+}
+
+export type NotificationSource = 'websocket' | 'push' | 'poll';
+
 export type NotificationEventCallback = (event: ZMAlarmEvent) => void;
 export type ConnectionStateCallback = (state: ConnectionState) => void;
